+++ lib/honeybadger/notice.rb
@@ -82,6 +82,9 @@ module Honeybadger
     # System stats
     attr_reader :stats
 
+    # The api_key to use when sending notice (optional)
+    attr_reader :api_key
+
     def initialize(args)
       self.args         = args
       self.exception    = args[:exception]
@@ -116,6 +119,7 @@ module Honeybadger
 
       self.hostname         = local_hostname
       self.stats            = Stats.all
+      self.api_key          = args[:api_key]
 
       self.source_extract_radius = args[:source_extract_radius] || 2
       self.source_extract        = extract_source_from_backtrace
@@ -143,6 +147,7 @@ module Honeybadger
     # Returns JSON representation of notice
     def as_json(options = {})
       {
+        :api_key => api_key,
         :notifier => {
           :name => notifier_name,
           :url => notifier_url,
@@ -231,7 +236,7 @@ module Honeybadger
       :environment_filters, :session_data, :project_root, :url, :ignore,
       :ignore_by_filters, :notifier_name, :notifier_url, :notifier_version,
       :component, :action, :cgi_data, :environment_name, :hostname, :stats, :context,
-      :source_extract, :source_extract_radius, :send_request_session
+      :source_extract, :source_extract_radius, :send_request_session, :api_key
 
     # Private: Arguments given in the initializer
     attr_accessor :args